- Reorganized the repository to make it easier to understand and maintain. - Added a new `base_xyz`` folder to the arctos_description package. - The `arctos_moveit_base_xyz`` package was also added to the repository.     - This package contains the MoveIt configuration files for the XYZ. - Removed `arctos_motor_driver_test`` package from the repository. - Added README.md file to all packages that were missing it.
The `arctos_bringup.launch.py`` file was modified to use the `base_xyz` URDF file while we focus on fixing software issues. It was also modified to wait for the controllers to be loaded before starting RVIZ and MoveIT. This allows the robot to first home all joints and set the zero position before starting RVIZ/MoveIT where the initial position should be 0.
The `ros2_controllers.yaml` from `arctos_moveit_base_xyz` has been modified with a lower update rate. We also added constraints to the controllers to allow the goal position to be reached in a reasonable time.